% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtotal_dupe_removal.R
\name{subtotal_dupe_removal}
\alias{subtotal_dupe_removal}
\title{Duplicate Subtotal Row Removal Function}
\usage{
subtotal_dupe_removal(
  data,
  column,
  iterator = 2,
  skip = 0,
  remove = "Total",
  lead_name = "Lead_Column"
)
}
\arguments{
\item{data}{Data report or data frame to remove duplicate row labels from.}

\item{column}{Column containing duplicate row labels.}

\item{iterator}{Minimum number of rows meant to be between each section. Usually two.}

\item{skip}{Number of rows to skip removing rows from. Usually zero. Can be used to avoid dealing with NA values or column labels.}

\item{remove}{Label of subtotals to be removed. Usually "Total".}

\item{lead_name}{Default name for lead column used to filter duplicates.}
}
\value{
The data report without duplicate subtotal rows.
}
\description{
Removes duplicate subtotal rows,
which may be created by totalling a variable with only one level.
In tables with many variables, some may have only one level within one section and many in other sections.
}
\details{
Adds a leading version of the requested column,
which places each row level in the same row as the row level two levels ahead.
This is usually the value 2 rows ahead of the current row,
although if not the iterator parameter can be used to account for this.
Then, if both the original and leading column equal the value to be removed,
then the row is considered a duplicate subtotal and is removed.
Note: Columns named Lead will receive a temporary suffix.
}
\examples{
library(dplyr)

group_by(mtcars, cyl, vs) \%>\% summarise(Sum_Wt = sum(wt), .groups = "keep") \%>\%
subtotal_row(mtcars, "wt") \%>\%
subtotal_dupe_removal(2)
group_by(mtcars, cyl, vs, am) \%>\% summarise(Mean_HP = mean(hp), .groups = "keep") \%>\%
subtotal_row(mtcars, "hp", "mean") \%>\%
subtotal_dupe_removal(3, skip = 1)
group_by(mtcars, cyl, vs, am) \%>\% summarise(Mean_HP = mean(hp), .groups = "keep") \%>\%
subtotal_row(mtcars, "hp", "mean") \%>\%
subtotal_dupe_removal(3, skip = 1)
}
