% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtotal_section.R
\name{subtotal_section}
\alias{subtotal_section}
\title{Subtotal Section Addition Function}
\usage{
subtotal_section(
  report,
  frame,
  vars = "Population",
  aggregator = "sum",
  exclude = numeric(0),
  agg_parameter = list(),
  subtotal_label = "Total"
)
}
\arguments{
\item{report}{A data report.}

\item{frame}{Data frame summarised by the data report.}

\item{vars}{Names of column(s) in the data frame which were aggregated in the data report.}

\item{aggregator}{Function to aggregate the data with.}

\item{exclude}{Vector of column indices determining which variables only require subtotal rows (as opposed to sections).}

\item{agg_parameter}{A named list of optional parameter(s) for the aggregation function to use.}

\item{subtotal_label}{Label to be used for subtotal rows.}
}
\value{
The grouped data report with subtotal sections included.
}
\description{
Adds subtotal sections to a data report. Designed to work with reports generated by group_by + summarise.
}
\details{
The dataset and report are factorized, and a series of reports are generated.
These reports are structured similarly to the original report - using the same grouping variables,
aggregating the given variables with the given aggregation function and optional parameters.
For best results, choose the same variables and aggregator used to generate the original report.
Each generated report has some grouping variables replaced by a chosen subtotal label,
effectively concentrating all levels of those variables
into one subtotal row for those variables - ultimately making up entire subtotal sections for some variables.
The subtotal reports are all combined with the original report,
and the combined report is grouped and sorted, sorting the subtotal label to the top for all variables.
}
\examples{
library(dplyr)

group_by(iris, Species) \%>\%
summarise(Sum_P_Len = sum(Petal.Length), .groups = "keep") \%>\%
subtotal_section(iris, vars = "Petal.Length")

group_by(mtcars, cyl, gear) \%>\%
summarise(Mean_Mpg = mean(mpg), .groups = "keep") \%>\%
subtotal_section(mtcars, vars = "mpg", aggregator = "mean")

group_by(mtcars, cyl, gear) \%>\%
summarise(Mean_Mpg = mean(mpg), Mean_Wt = mean(wt), .groups = "keep") \%>\%
subtotal_section(mtcars, vars = c("mpg", "wt"),
aggregator = "mean", exclude = 1:2)

group_by(iris, Species, Petal.Width) \%>\%
summarise(Max_S_Width = max(Sepal.Width), Max_S_Length = max(Sepal.Length), .groups = "keep") \%>\%
subtotal_section(iris, vars = c("Sepal.Width", "Sepal.Length"),
aggregator = "max", agg_parameter = list(na.rm = TRUE))

group_by(mtcars, qsec, carb, hp, gear) \%>\%
summarise(Min_Cyl = min(cyl), Min_Drat = min(drat), Min_Wt = min(wt), .groups = "keep") \%>\%
subtotal_section(mtcars, vars = c("cyl", "drat", "wt"),
aggregator = "min", agg_parameter = list(na.rm = TRUE),
subtotal_label = "Cars_Total", exclude = c(1, 4))
}
